[fix] Google 로그인 로직 수정
- 필요한 정보만을 LoginResponse로 앱에 전송하도록 수정
- 로그인 시 소셜 플랫폼 이름, Id 토큰, FCM 토큰 전부 전송하도록 수정
- 또한 넥슨 Open API와 통신할 때에는 캐릭터 목록을 조회하는 API를 제외하고는 전부 서버에 등록된 API Key를 헤더에 포함하도록 수정
 - 유저의 캐릭터 목록을 조회할 때에만 유저가 입력했던 API Key를 헤더에 포함

+++ src/main/kotlin/com/sixclassguys/maplecalendar/infrastructure/external/dto/AccountCharacter.kt
@@ -16,5 +16,5 @@ data class AccountCharacter(
     val characterClass: String,
 
     @JsonProperty("character_level")
-    val characterLevel: Int
+    val characterLevel: Long
 )
